Add CLI to generate reference docs
Copied from Epinio.
Importing 'main' as a package is not possible, so extracting the cobra
command generation into separate packages, like 'fleet cli' already did.
